fix deno size error
